+++ update.sh
@@ -8,30 +8,47 @@ fi
 sudo apt update
 sudo apt install jq -y
 
-whiptail --title "Pihole Speedtest Mod Updater" --msgbox "Pihole Speedtest Mod Updater. \nSupport : https://github.com/arevindh/pihole-speedtest " 8 78
-if (whiptail --title "Pihole Speedtest Mod Updater" --yesno "Proceed to update ?" 8 78); then
+whiptail --title "Pi-hole Speedtest Mod Updater" --msgbox "Pi-hole Speedtest Mod Updater. \nSupport : https://github.com/arevindh/pihole-speedtest " 8 78
+if (whiptail --title "Pi-hole Speedtest Mod Updater" --yesno "Proceed to update ?" 8 78); then
     echo "Proceeding with update"
 else
-    exit 1;
+    exit 1
 fi
 
-echo "Updating Pihole and Speedtest-mod"
-
-#Revert admin to Pihole's newest
+echo "Reverting files..."
 cd /var/www/html
-rm -rf admin/
-git clone https://github.com/pi-hole/AdminLTE admin
+# if org_admin exists, mv it to admin, else clone admin from URL
+if [ -d /var/www/html/org_admin ]; then
+    rm -rf admin
+    mv org_admin admin
+else
+    rm -rf admin
+    git clone https://github.com/pi-hole/AdminLTE admin
+fi
 
-#Revert pihole's webpage
 cd /opt/pihole/
-rm webpage.sh
-wget https://github.com/pi-hole/pi-hole/raw/master/advanced/Scripts/webpage.sh
-chmod +x webpage.sh
+if [ -f /opt/pihole/webpage.sh.org ]; then
+    mv webpage.sh.org webpage.sh
+else
+    wget https://github.com/pi-hole/pi-hole/raw/master/advanced/Scripts/webpage.sh
+    chmod +x webpage.sh
+fi
+if [ -f /opt/pihole/version.sh.org ]; then
+    mv version.sh.org version.sh
+else
+    wget https://github.com/pi-hole/pi-hole/raw/master/advanced/Scripts/version.sh
+    chmod +x version.sh
+fi
+echo "Files reverted."
 
-#Update Pihole
 pihole -up
 
-#Update lastest speedtest-mod
+if [ -n "$1" ] && [ "$1" = "un" ]; then
+    echo "Speedtest Mod Uninstall Complete"
+    exit 0
+fi
+
+echo "Updating Speedtest Mod..."
 cd /var/www/html
 rm -rf pihole_admin
 rm -rf admin_bak
@@ -46,12 +63,11 @@ wget https://github.com/arevindh/pi-hole/raw/master/advanced/Scripts/webpage.sh
 chmod +x webpage.sh
 
 mv version.sh version.sh.org
-wget https://raw.githubusercontent.com/arevindh/pi-hole/master/advanced/Scripts/version.sh
+wget https://github.com/arevindh/pi-hole/raw/master/advanced/Scripts/version.sh
 chmod +x version.sh
 
 #Update version info
 pihole updatechecker local
-
 
 whiptail --title "Pihole Speedtest Mod" --msgbox "Update complete" 8 78
 exit 0
